+++ src/Graphics/Tests/ShadowedPRT.cpp
@@ -177,7 +177,7 @@ void ShadowedPRT::OnInitialize()
 					&m_Vertices[vtxCount], &m_Indices[idxCount],
 
 					mnfVtxCount, sizeof(VertexP), 0.05f,
-					(s32)floral::geo_vertex_format_e::position,
+					floral::geo_vertex_format_e::position,
 					&m_MnfVertices[mnfVtxCount], &m_MnfIndices[mnfIdxCount]);
 
 			for (size i = vtxCount; i < vtxCount + genResult.vertices_generated; i++)
@@ -204,7 +204,7 @@ void ShadowedPRT::OnInitialize()
 					&m_Vertices[vtxCount], &m_Indices[idxCount],
 
 					mnfVtxCount, sizeof(VertexP), 0.05f,
-					(s32)floral::geo_vertex_format_e::position,
+					floral::geo_vertex_format_e::position,
 					&m_MnfVertices[mnfVtxCount], &m_MnfIndices[mnfIdxCount]);
 
 			for (size i = vtxCount; i < vtxCount + genResult.vertices_generated; i++)
@@ -230,7 +230,7 @@ void ShadowedPRT::OnInitialize()
 					&m_Vertices[vtxCount], &m_Indices[idxCount],
 
 					mnfVtxCount, sizeof(VertexP), 0.05f,
-					(s32)floral::geo_vertex_format_e::position,
+					floral::geo_vertex_format_e::position,
 					&m_MnfVertices[mnfVtxCount], &m_MnfIndices[mnfIdxCount]);
 
 			for (size i = vtxCount; i < vtxCount + genResult.vertices_generated; i++)
@@ -255,7 +255,7 @@ void ShadowedPRT::OnInitialize()
 					&m_Vertices[vtxCount], &m_Indices[idxCount],
 
 					mnfVtxCount, sizeof(VertexP), 0.05f,
-					(s32)floral::geo_vertex_format_e::position,
+					floral::geo_vertex_format_e::position,
 					&m_MnfVertices[mnfVtxCount], &m_MnfIndices[mnfIdxCount]);
 
 			for (size i = vtxCount; i < vtxCount + genResult.vertices_generated; i++)
